Commit(Claire Finn)
I added an RSS feed Icon, and a link to an RSS Feed to each page. The
RSS will not display as its not linked to a published website but the
code is there. The RSS feeds on each page link to the same article for
now but I will change that for the next commit.