RSS Feeds(Claire Finn)
Added an RSS Icon and a link to an RSS Feed to each page. The RSS feed will not currently display in the browser as it's normally attached to a URL of a published website (which ours isn't). The code for the RSS is there though showing the news items we would have liked to have displayed on each page. At the moment, all of the links are linking to the same feed but I will add some info specific to each page in the next commit.